+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/availability/PermissiveAvailability.java
@@ -60,4 +60,9 @@ public class PermissiveAvailability extends ConcreteAvailability {
                 .flatMap(List::stream)
                 .collect(SimplifiedIntervalList.getCollector());
     }
+
+    @Override
+    public String toString() {
+        return "permissive availability";
+    }
 }