+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Current
 - [Add Reciprocal `satisfies()` relationship complementing `satisfiedBy()` on Granularity](https://github.com/yahoo/fili/issues/222)
 
 - [MetricUnionAvailability](https://github.com/yahoo/fili/pull/193)
-    * Added `MetricUnionAvailability`  which joins physical tables with the same dimension schema
+    * Added `MetricUnionAvailability` which joins physical tables with the same dimension schema
     but non-intersecting metric schemas. Metrics on the `MetricUnionAvailability` are sourced from exactly one of the participating tables.
 
 - [Method for finding coarsest ZonedTimeGrain](https://github.com/yahoo/fili/pull/230)